Add application for homework by using Cpp in QT
Application consist two task. First task is creating class counter in C++ with function increase by one and clear value of counter. It is used in application with two buttons carring function increase and clear of class Counter and TextField displaying value of Counter. Second task is creating class list of strings with method append string in the end of list and remove last item of list. This class is used in application with two buttons first of these append string entered in InputTextField and second of these removed last item of list.